feat: allow TYPEORM_SYNC env flag to create schema in production

+++ src/authors/authors.service.ts
@@ -4,6 +4,7 @@ import { Repository } from 'typeorm';
 import { AutherInterface } from './interfaces/auther.interface';
 import { Author } from './entities/author.entity';
 import { AuthorDto } from './dto/author.dto';
+import { ResponseService } from 'src/common/services/response.service';
 
 @Injectable()
 export class AuthorsService {
@@ -13,24 +14,48 @@ export class AuthorsService {
         private readonly authorRepository: Repository<Author>,
     ) { }
 
-    create(autor: AutherInterface) {
+    async create(autor: AuthorDto) {
         const auteurSave = this.authorRepository.create(autor)
-        return this.authorRepository.save({
+        const savedAuthor = await this.authorRepository.save({
             ...auteurSave
         });
+        return ResponseService.created(savedAuthor, 'Auteur créé avec succès');
     }
 
 
-    update(id: string, authorDto: AuthorDto) {
-        return this.authorRepository.update(id, authorDto);
+    async update(id: string, authorDto: AuthorDto) {
+
+        const exists = await this.checkAuthorExists(parseInt(id));
+        if (!exists) {
+            return ResponseService.notFound('Auteur non trouvé');
+        }
+        const updatedAuthor = await this.authorRepository.update(id, authorDto);
+        return ResponseService.updated(updatedAuthor, 'Auteur modifié avec succès');
     }
-    remove(id: string) {
-        return this.authorRepository.delete(id);
+    
+    async remove(id: string) {
+        const exists = await this.checkAuthorExists(parseInt(id));
+        if (!exists) {
+            return ResponseService.notFound('Auteur non trouvé');
+        }
+        await this.authorRepository.delete(id);
+        return ResponseService.deleted('Auteur supprimé avec succès');
     }
-    findAll() {
+    
+    async findAll() {
         return this.authorRepository.find();
     }
-    findOne(id: string) {
+    
+    async findOne(id: string) {
+        const exists = await this.checkAuthorExists(parseInt(id));
+        if (!exists) {
+            return ResponseService.notFound('Auteur non trouvé');
+        }
         return this.authorRepository.findOne({ where: { id: parseInt(id) } });
+    }
+    
+    // check l'existence d'un auteur
+    async checkAuthorExists(id: number): Promise<boolean> {
+        return this.authorRepository.exists({ where: { id } });
     }
 }
